+++ src/modules/auth.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+import { AuthController } from '../api/controllers/auth.controller';
+import { AuthService } from '../api/services/auth.service';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthGuard } from '../security/auth.guard';
